+++ src/main/java/xyz/jungha/job/command/exp/SubtractExpCommand.java
@@ -28,7 +28,7 @@ public class SubtractExpCommand extends ExpSubCommand {
     }
 
     @Override
-    protected boolean isValidAmount(int amount, CommandSender sender) {
+    protected boolean isValidAmount(double amount, CommandSender sender) {
         if (amount <= 0) {
             sender.sendMessage(MINI_MESSAGE.deserialize("[<gold>직업<white>] <red>감소량은 1 이상이어야 합니다."));
             return false;
@@ -37,12 +37,12 @@ public class SubtractExpCommand extends ExpSubCommand {
     }
 
     @Override
-    protected void applyChange(OfflinePlayer player, Jobs job, int amount) {
+    protected void applyChange(OfflinePlayer player, Jobs job, double amount) {
         Job.getInstance().getJobRepo().subtractJobExp(player, job, amount);
     }
 
     @Override
-    protected String getSuccessMessage(OfflinePlayer player, Jobs job, int amount) {
+    protected String getSuccessMessage(OfflinePlayer player, Jobs job, double amount) {
         return String.format("[<gold>직업<white>] <green>%s</green>님의 <yellow>%s</yellow> 직업 경험치를 <gold>-%d</gold> 만큼 감소시켰습니다.", player.getName(), job.getDisplayName(), amount);
     }
 }
